backend/requirements.txt
python-docx==0.8.11
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
